+++ app/src/main/java/com/example/myapplication_1/Fragments/ServiceFragment.java
@@ -66,13 +66,13 @@ public class ServiceFragment extends Fragment {
         relativeLayout2.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String message = textView.getText().toString();
-                Bundle _args = new Bundle();
-                _args.putString("message", message);
+//                String message = textView.getText().toString();
+//                Bundle _args = new Bundle();
+//                _args.putString("message", message);
+//                programErrorTwoFragment.setArguments(_args);
                 FragmentTransaction fragmentTransaction = getFragmentManager().beginTransaction();
                 fragmentTransaction.replace(R.id.rl_main, programErrorTwoFragment);
                 fragmentTransaction.commit();
-                programErrorTwoFragment.setArguments(_args);
             }
         });
         relativeLayout3.setOnClickListener(new View.OnClickListener() {
